REFACTOR: move GraphBuilder to its own file
