+++ src/graph/index.ts
@@ -1,6 +1,7 @@
 export * from './edge';
 export * from './forward_rule';
 export * from './graph';
+export * from './graph_builder';
 export * from './loader';
 export * from './node';
 export * from './types';
